% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/terminal.R
\name{asc2chr}
\alias{asc2chr}
\title{Convert ASCII Code to Character}
\usage{
asc2chr(x)
}
\arguments{
\item{x}{ASCII code to convert}
}
\description{
Convert ASCII code to character.
}
\examples{
asc2chr(120)
}
\references{
See \url{http://datadebrief.blogspot.com/2011/03/ascii-code-table-in-r.html}
}

