% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro_scoring.R
\name{av_cell_dist}
\alias{av_cell_dist}
\title{Computes a distance-based score for each MST}
\usage{
av_cell_dist(cell_MST, clusterLabels, points, time)
}
\arguments{
\item{cell_MST}{List returned by \code{create_dMST()} containing MST-specific information, such as
MST structure, starting/terminal nodes, etc.}

\item{clusterLabels}{A membership vector for clustering the coordinates into groups.}

\item{points}{Coordinates for each cell. For low-dimensional projections from PCA, these coordinates use
the first 3 PC's and will have no time contribution. Note that if UMAP is used, there will be a time contribution as the first
three dimensions are UMAP1, UMAP2, and the weighted time vector.}

\item{time}{Vector of sampling time points corresponding to each cell.}
}
\value{
List containing the (1) number of lineages, (2) score, and a (3) binary membership matrix indicating
which lineage the cell is closest to. This membership matrix will be used for detecting global lineages based on
all MSTs.
}
\description{
This function projects each cell to each lineage of the MST to gauge how well the
MST covers the low-dimensional projection of data. It includes penalties based on whether all of the
nodes are used, and if the initial and terminal nodes are consistent with the experimental sampling time.
}
