% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ballparks}
\alias{ballparks}
\title{Ballpark codes and information}
\format{Data frame with columns
\describe{
\item{parkid}{Ballpark ID}
\item{name}{Name of the ballpark}
\item{aka}{Alternative name for the ballpark}
\item{city}{City in which the ballpark is located}
\item{state}{State in which the ballpark is located}
\item{start}{Date the ballpark came into service}
\item{end}{Date of the ballpark's last service}
\item{league}{League in which the ballpark (primarily) served}
\item{notes}{additional information}
}}
\source{
\url{http://www.retrosheet.org/parkcode.txt}
}
\usage{
ballparks
}
\description{
Ballpark codes and information
}
\keyword{datasets}
