% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/terminal.R
\name{chr2asc}
\alias{chr2asc}
\title{Convert Character to ASCII Code}
\usage{
chr2asc(x)
}
\arguments{
\item{x}{character to convert}
}
\description{
Convert character to ASCII code.
}
\examples{
chr2asc("x")
}
\references{
See \url{http://datadebrief.blogspot.com/2011/03/ascii-code-table-in-r.html}
}

