% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/terminal.R
\name{colorize}
\alias{colorize}
\title{Colorize Text}
\usage{
colorize(s, fc = NA, bc = NA, force = FALSE)
}
\arguments{
\item{s}{text to colorize}

\item{fc}{foreground color constant. Defaults to \code{NA} which results in the terminal default.}

\item{bc}{background color constant. Defaults to \code{NA} which results in the terminal default.}

\item{force}{if \code{TRUE}, adds color codes even on Windows (see e.g. ConEmu or ansicon). Defaults to \code{FALSE}.}
}
\description{
Colorize text with ANSI escape codes for the *nix terminal.
}
\examples{
cat(colorize(0:7, fc = x_colors$black, bc = 0:7))
}

