% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro_distance.R
\name{create_dMST}
\alias{create_dMST}
\title{Calculating the minimum spanning tree for trajectory inference}
\usage{
create_dMST(retro_obj, kmedoids)
}
\arguments{
\item{retro_obj}{RETRO meta object}

\item{kmedoids}{Cluster centroids and labels obtained from \code{kmnn_cluster()}.}
}
\value{
List of MST trajectory information: minimum spanning tree object (MST),
starting and terminal nodes (Start, Terminal), constituent lineages inferred (Lineages),
cluster labels for cell coordinates (clusterLabels), node IDs (ID),
MST edge cutoff based on time (Dt), and a logical value for whether the cells were reclustered (Reclus)
}
\description{
This function uses RETRO functions \code{recluster()} and \code{get_cluster_time()} to figure out
clustering and establish the initial and final nodes of the minimum spanning tree. It
depends on functions from the \code{igraph} package for most graph-based calculations.
}
