% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro_lineages.R
\name{create_meta_obj}
\alias{create_meta_obj}
\title{Create RETRO object}
\usage{
create_meta_obj(
  obj,
  coordinates,
  terminal_cells = NULL,
  starting_cells = NULL,
  start = NULL,
  period = NULL,
  k,
  clustering,
  scoring
)
}
\arguments{
\item{coordinates}{Matrix of cell coordinates with a time contribution
obtained using the \code{weight_coord()} function.}

\item{terminal_cells}{(optional) List of cells belonging to the terminal ends of the trajectory.
Each list element can contain the cells corresponding to one terminal cell type.}

\item{starting_cells}{(optional) Vector of cells belonging to the starting end of the trajectory.
RETRO currently allows for only one starting node.}

\item{start}{Character parameter that is either "Average" or "Mode." If "Average,"
the starting node is assigned to the cluster with the lowest average sampling time.
If "Mode," the starting node is assigned the cluster with the greatest representation of the lowest sampling time-point.
"Mode" is recommended for determining the starting node.}

\item{period}{Specifies the minimum difference in time where one can expect the cells
to return to an earlier state gene expression (make a cycle).}

\item{k}{Range of clusters used for RETRO pseudotime analysis.}

\item{clustering}{List of clusterings applied to the data obtained the function \code{scoring()} that
is organized by the corresponding K clusters used. The length of the list should be
number of K clusters used, \code{length(k)}, and each list element should be the length \code{num_scores}.}

\item{scoring}{List of MSTs obtained from the function \code{scoring()} that
is organized by the corresponding K clusters used. The length of the list should be
number of K clusters used, \code{length(k)}, and each list element should be the length \code{num_scores}.
The list contains the scores and structures of each MST determined.}
}
\value{
RETRO object.
}
\description{
This function stores all necessary information to carry out analysis with RETRO.
}
