% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.R
\name{data.frame0}
\alias{data.frame0}
\title{Data Frames}
\usage{
data.frame0(...)
}
\arguments{
\item{...}{passed to \code{data.frame} with \code{stringsAsFactors = FALSE}}
}
\description{
Create a data frame with \code{stringsAsFactors = FALSE}.
}
\keyword{internal}

