% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/draw.R
\name{draw}
\alias{draw}
\title{Draw}
\usage{
draw(d, ch = "o", fc = NA, bc = NA, h = 28, t = 0.01, text = "")
}
\arguments{
\item{d}{data frame to draw. Columns required: \code{x} and \code{y} with coordinates (top left is (0,0), \code{x} increases downwards, \code{y} increases rightwards). Warning: duplicated rows (regarding \code{x} and \code{y}) are excluded. Columns optional: \code{ch} with the characters to draw, \code{fc} and \code{bc} with the foreground and background color constants.}

\item{ch}{the character to use if not given in the data frame. Defaults to \code{"o"}.}

\item{fc}{the foreground color constant to use if not given in the data frame. Defaults to \code{NA} which results in the terminal default.}

\item{bc}{the background color constant to use if not given in the data frame. Defaults to \code{NA} which results in the terminal default.}

\item{h}{the height of the space where the drawing is placed. Defaults to \code{28}.}

\item{t}{seconds to wait after drawing. Defaults to \code{0.01}.}

\item{text}{text to write at the end of the drawing. Defaults to empty string.}
}
\description{
Draw on the terminal. The source is a coordinated data frame.
}
\examples{
draw(data.frame(x = c(1:20, 20:1), y = c(1:20, 1:20)))
draw(coordinate(retro_example("hello_world")))
}

