% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/draw.R
\name{draw_effect}
\alias{draw_effect}
\title{Draw with Effects}
\usage{
draw_effect(file, coords = coordinate(file), effect = "default", w = 20,
  t = 0.01)
}
\arguments{
\item{file}{input file}

\item{coords}{input coordinates. Defaults to \code{coordinate(file)}.}

\item{effect}{effect to apply. Possible values: \code{"default"}, \code{"random_order"}, \code{"from_right"}.}

\item{w}{width of drawing used with effect \code{"from_right"}. Defaults to \code{20}.}

\item{t}{seconds between frames, passed to function \code{draw}. Defaults to \code{0.01}.}
}
\description{
Draw with effects on the terminal.
}
\examples{
\dontrun{draw_effect(retro_example('hello_world'), effect = "from_right")}
}

