% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro.R
\name{etl_extract.etl_retro}
\alias{etl_extract.etl_retro}
\alias{etl_transform.etl_retro}
\alias{etl_load.etl_retro}
\title{Download Retrosheet data}
\usage{
\method{etl_extract}{etl_retro}(obj, season, ...)

\method{etl_transform}{etl_retro}(obj, season, ...)

\method{etl_load}{etl_retro}(obj, season, ...)
}
\arguments{
\item{obj}{the ETL object}

\item{season}{the season}

\item{...}{currently ignored}
}
\description{
Download Retrosheet data
}
\examples{
\dontrun{
  system("mysql -e 'CREATE DATABASE IF NOT EXISTS retrosheet;'")
  db <- src_mysql_cnf("retrosheet", hostname = "127.0.0.1")
  retro <- etl("retro", db = db, dir = "~/dumps/retro/")

  retro \%>\%
    etl_extract(season = 2013:2016) \%>\%
    etl_transform(season = 2014:2016) \%>\%
    etl_load(season = 2014)
}
}
