% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{event_codes}
\alias{event_codes}
\title{Event codes}
\format{Data frame with two columns
\describe{
\item{EVENT_CD}{Event code}
\item{EVENT}{Description of the event}
}}
\source{
\url{https://github.com/chadwickbureau/chadwick/blob/master/doc/cwevent.rst}
}
\usage{
event_codes
}
\description{
Event codes
}
\keyword{datasets}
