% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro_clustering.R
\name{find_nearest_cell}
\alias{find_nearest_cell}
\title{Clustering functions in RETRO}
\usage{
find_nearest_cell(centroid, coordinates)
}
\arguments{
\item{centroid}{List of K-means centroids of the coordinate clusters.}

\item{coordinates}{Matrix of cell coordinates with a time contribution
obtained using the \code{weight_coord()} function.}
}
\value{
The index of the closest cell to each centroid.
}
\description{
This function finds the cell that is closest to the centroid of its cluster.
}
