% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/draw.R
\name{generate_border}
\alias{generate_border}
\title{Generate Border}
\usage{
generate_border(h = 27, w = 52, fc = NA, bc = NA)
}
\arguments{
\item{h}{height of the border. Defaults to \code{27}.}

\item{w}{width of the border. Defaults to \code{52}.}

\item{fc}{foreground color constant. Defaults to \code{NA} which results in the terminal default.}

\item{bc}{background color constant. Defaults to \code{NA} which results in the terminal default.}
}
\description{
Generates a border (rectangle) as a coordinated data frame.
}
\examples{
generate_border()
}

