% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/maze.R
\name{generate_maze}
\alias{generate_maze}
\title{Generate a random basic maze}
\usage{
generate_maze(w = 25, h = 25)
}
\arguments{
\item{w}{width}

\item{h}{height}
}
\value{
\code{data.frame} ready to be passed to \code{draw}
}
\description{
Generate a random basic maze
}
\examples{
draw(generate_maze())
draw(generate_maze(10, 10))
draw(generate_maze(80, 25))
}
\references{
\url{http://10print.org/}
}

