% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro_scoring.R
\name{generate_paths}
\alias{generate_paths}
\title{Infer and score MST}
\usage{
generate_paths(retro_obj, kmedoids, max_k)
}
\arguments{
\item{retro_obj}{RETRO meta object.}

\item{kmedoids}{Cluster centroids and labels obtained from \code{kmnn_cluster()}.}

\item{max_k}{Maximum number of clusters from the user-specified range of K.}
}
\value{
List containing the (1) MST scoring information from \code{av_cell_dist()}
and (2) clustering.
}
\description{
This function is a wrapper for key RETRO functions, namely
\code{create_dMST()} for MST inference and \code{av_cell_dist()} for MST scoring.
It is iteratively called in the function \code{scoring()} to generate MSTs for
coordinates clustered in a variety of ways.
}
