% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/terminal.R
\name{get_ch}
\alias{get_ch}
\title{Non-Blocking Keyboard Input}
\usage{
get_ch(n = 50)
}
\arguments{
\item{n}{number of loops to wait at most. Defaults to \code{50} (which is 500 ms).}
}
\value{
Returns the ASCII code of the key pressed, or 0 if there was no key pressed.
}
\description{
Non-blocking keyboard input with 10 ms delay. Works only on the terminal (and not on the console).
}
\examples{
get_ch()
}
\references{
See \url{http://stackoverflow.com/a/6665815} and \url{http://stackoverflow.com/a/13129698}
}

