% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro_distance.R
\name{get_cluster_time}
\alias{get_cluster_time}
\title{Determing time information per cluster}
\usage{
get_cluster_time(
  clusterLabels,
  time,
  terminal_cells = NULL,
  starting_cells = NULL,
  threshold = 0.1,
  max_k
)
}
\arguments{
\item{clusterLabels}{A membership vector for clustering the coordinates into groups.}

\item{time}{Vector of sampling time points corresponding to each cell.}

\item{terminal_cells}{(optional) List of cells belonging to the terminal ends of the trajectory.
Each list element can contain the cells corresponding to one terminal cell type.}

\item{starting_cells}{(optional) Vector of cells belonging to the starting end of the trajectory.
RETRO currently allows for only one starting node.}

\item{threshold}{(optional) Cutoff to classify whether clusters are terminal ends of the trajectory. Default is 0.1.
Note that this cutoff is scaled by number of clusters (K) / maximum number of clusters specified (Max_K).}

\item{max_k}{Maximum number of clusters from the user-specified range of K.}
}
\value{
Returns consensus information about clusters such as their average and mode time (AvTime, Mode),
as well as the inferred initial and terminal clusters based on sampling time (Terminal_ID, Start_ID).
}
\description{
This function infers the starting and terminal clusters based on time information.
Such information can also be user-specified if preferred.
}
