% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro_pseudotime_est.R
\name{get_mapped_cells}
\alias{get_mapped_cells}
\title{Final cell mapping}
\usage{
get_mapped_cells(retro_obj)
}
\arguments{
\item{retro_obj}{MST-specific object from RETRO containing information about the dataset
and the optimal MST used in curve fitting.}
}
\value{
Updated elements in \code{retro_obj} to store RETRO pseudotime in \code{pseudotime}
and intermediate fitting processes in \code{pseudotime_graph}.
}
\description{
This function pdates lineage membership information
such that every cell maps to a branch of the optimal MST.
}
