% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chadwick.R
\name{install_chadwick}
\alias{install_chadwick}
\alias{check_chadwick}
\alias{find_chadwick}
\alias{set_library_path}
\title{Install Chadwick}
\usage{
install_chadwick(obj, ...)

check_chadwick()

find_chadwick()

set_library_path()
}
\arguments{
\item{obj}{An \code{etl_retro} object}

\item{...}{currectly ignored}
}
\description{
Install Chadwick
}
