% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro_clustering.R
\name{kmnn_cluster}
\alias{kmnn_cluster}
\title{Clustering strategy adapted from K-means.}
\usage{
kmnn_cluster(coordinates, num_centers)
}
\arguments{
\item{coordinates}{Matrix of cell coordinates with a time contribution
obtained using the \code{weight_coord()} function.}

\item{num_centers}{Determines number of clusters.}
}
\value{
A list containing the new cluster centroids \code{centers} and
a vector of labels \code{clusterLabels} grouping the coordinates.
}
\description{
This function is an adaptation of k-means that maps K-means centroids
in each cluster to the nearest cell from the original data.
}
