% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.R
\name{paste1}
\alias{paste1}
\title{Concatenate Strings}
\usage{
paste1(...)
}
\arguments{
\item{...}{passed to \code{paste0} with \code{collapse = ""}}
}
\description{
Concatenate vectors into a single string.
}
\keyword{internal}

