% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro_pseudotime_est.R
\name{pseudotime_fit}
\alias{pseudotime_fit}
\title{Pseudotime fitting function}
\usage{
pseudotime_fit(retro_pt_obj, retro_obj)
}
\arguments{
\item{retro_pt_obj}{Pseudotime-specific object from RETRO which contains consensus information
across MSTs regarding the global number of lineages and cell membership to each lineage.}

\item{retro_obj}{MST-specific object from RETRO containing information about the dataset
and the optimal MST used in curve fitting.}
}
\value{
Updated list for \code{retro_pt_obj} where \code{retro_pt_obj[["Pseudotime"]]} stores RETRO pseudotime.
Remaining arguments "Pseudotime_Mat", "Graph_List," and "Fitting" store intermediate steps from the
pseudotime inference process.
}
\description{
This function uses lineage information from \code{retro_pt_obj} and
MST-specific information from \code{retro_obj} to infer pseudotime values.
}
