% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{psupertime_density}
\alias{psupertime_density}
\title{Pseudotime density plot}
\usage{
psupertime_density(time, pseudotime, bw = 0.5)
}
\arguments{
\item{time}{Vector of experimental sampling time assigned to each cell.}

\item{pseudotime}{Vector of pseudotime estimates assigned to each cell.}

\item{bw}{(optional) Density plot bandwidth. Default at 0.5.}
}
\value{
A ggplot2 object showing the normalized density of pseudotime values grouped by sampling time.
}
\description{
This function creates a density plot of pseudotime that is grouped
by discrete experimental time labels. This plot gives insight into whether the inferred
pseudotime is consistent with the progression of real time.
}
