% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro_class_def.R
\docType{class}
\name{RETRO-class}
\alias{RETRO-class}
\alias{set_RETRO_class}
\alias{RETRO}
\title{RETRO: S4 class for pseudotime analysis.}
\value{
An object of class \code{RETRO}
}
\description{
This S4 class contains slots for containing data and parameters necessary
for analysis with RETRO functions, and also stores the final inferred pseudotime.
}
\section{Slots}{

\describe{
\item{\code{coordinates}}{Coordinates determined by \code{weight_coord()} that include a contribution from sampling time.}

\item{\code{coordinates_uw}}{Coordinates determined by \code{weight_coord()} that have an \strong{optional} contribution from sampling time.
It is recommended that these coordinates are un-weighted so that the initial clustering covers the entire low-dimensional projection.}

\item{\code{time}}{Numeric vector of sampling time points corresponding to each cell.}

\item{\code{k_range}}{Numeric vector for the range of K clusters that will be used to partition the coordinates}

\item{\code{start}}{(optional) Character that is either "Average" or "Mode" to determine the starting node (if no starting cells are specified).
If "Average," the starting node is assigned to the cluster with the lowest average sampling time.
If "Mode," the starting node is assigned the cluster with the greatest representation of the lowest sampling time-point.
Default is "Mode".}

\item{\code{terminal_cells}}{(optional) List of cells belonging to the terminal ends of the trajectory.
Each list element can contain the cells corresponding to one terminal cell type.}

\item{\code{starting_cells}}{(optional) Vector of cells belonging to the starting end of the trajectory.
RETRO currently allows for only one starting node.}

\item{\code{threshold}}{(optional) Cutoff to classify whether clusters are terminal ends of the trajectory. Default is 0.1.
Note that this cutoff is scaled by number of clusters (K) / maximum number of clusters specified (Max_K).}

\item{\code{period}}{(optional) Specifies the minimum difference in time where one can expect the cells
to return to an earlier state gene expression (make a cycle).}

\item{\code{all_k}}{List containing cluster labels and centroids for each iteration of clustering on the range of K}

\item{\code{all_scores}}{List containing MST scores and cell membership for all K clusters in the specified range}

\item{\code{num_lineages}}{!!! include description here}

\item{\code{centroids}}{!!!! include description here}

\item{\code{lin_membership}}{!!!!!!!description goes here}

\item{\code{cells_to_lin}}{!!!!!!description goes here}

\item{\code{RETRO_MST}}{!!!!!!!description goes here}

\item{\code{RETRO_Curve}}{!!!!!!description goes here}

\item{\code{pseudotime_graph}}{!!! include description here}

\item{\code{pseudotime}}{!!!! include description here}
}}

