% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/helpers.R
\name{retro_example}
\alias{retro_example}
\title{Path to Package Examples}
\usage{
retro_example(file)
}
\arguments{
\item{file}{filename of the example without ".txt"}
}
\description{
Returns the path to \code{retro} example text files.
}
\examples{
retro_example("hello_world")
}

