% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scd_bifurcation.R
\docType{data}
\name{scd_bifurcation}
\alias{scd_bifurcation}
\title{Simulated bifurcation data for RETRO analysis}
\format{
A list containing 3 elements:
\describe{
\item{Counts}{Log-transformed normalized gene expression matrix containing 1000 cells and 500 genes.}
\item{Time}{Ground truth sampling time.}
\item{Time_Cont}{Ground truth continuous developmental time used for benchmarking pseudotime.}
}
}
\usage{
scd_bifurcation
}
\description{
This dataset contains count data modeling a cellular bifurcation.
}
\keyword{datasets}
