% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro_scoring.R
\name{scoring}
\alias{scoring}
\title{Parallelization of MST generation and scoring}
\usage{
scoring(
  coordinates,
  coordinates_uw,
  time,
  k,
  num_scores,
  num_cells,
  terminal_cells = NULL,
  starting_cells = NULL,
  max_k,
  start = "Average",
  threshold = 0.1,
  period = period
)
}
\arguments{
\item{coordinates}{Coordinates determined by \code{weight_coord()} that include a contribution from sampling time.}

\item{coordinates_uw}{Coordinates determined by \code{weight_coord()} that have an \strong{optional} contribution from sampling time.
It is recommended that these coordinates are un-weighted so that the initial clustering covers the entire low-dimensiona projection.}
}
\value{
List containing the (1) MST scoring infomration from \code{av_cell_dist()}
and (2) clustering.
}
\description{
This function is a wrapper for the \code{generate_paths()} function
to parallelize iterative MST generation and scoring across the range of specified K clusters.
It is recommended to run \code{scoring()} in an lapply() to contain all
iterative clustering and scoring data in one .rda file.
}
