% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro_distance.R
\name{weight_coord}
\alias{weight_coord}
\title{Defining cell coordinates using gene expression and time}
\usage{
weight_coord(eset, weight)
}
\arguments{
\item{eset}{Biobase ExpressionSet object that contains dimensionality reduction \code{experimentData(eset)@other[[1]]}
and sampling time information \code{phenoData(scd_bifurcation)@data[["time"]]}.}

\item{weight}{A numerical value indicating contribution from PCA for scaling the \code{time} vector.
Default value is 1 (where PCA and time have an equal contribution). Increasing lambda will increase contribution from PCA.}
}
\value{
Returns the updated cell coordinates based on the PCA projection and sampling time.
}
\description{
This function is a wrapper for the function \code{pc_distance_function()}.
}
