% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/terminal.R
\docType{data}
\name{x_colors}
\alias{x_colors}
\title{Terminal Color Constants}
\format{\preformatted{List of 8
 $ black  : num 0
 $ red    : num 1
 $ green  : num 2
 $ yellow : num 3
 $ blue   : num 4
 $ magenta: num 5
 $ cyan   : num 6
 $ white  : num 7
}}
\usage{
x_colors
}
\description{
Terminal color constants.
}
\examples{
x_colors$red
}
\keyword{datasets}

