% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{boxplot_scoring}
\alias{boxplot_scoring}
\title{Boxplots to visualize score distribution}
\usage{
boxplot_scoring(all_scores, num_scores, k)
}
\arguments{
\item{all_scores}{List of MSTs obtained from the function \code{scoring()} that
is organized by the corresponding K clusters used. The length of the list should be
number of K clusters used, \code{length(k)}, and each list element should be the length \code{num_scores}.
The list contains the scores and structures of each MST determined.}

\item{num_scores}{Number of MSTs (scored MSTs) per clustering. Default is 100.}

\item{k}{Range of clusters used for RETRO pseudotime analysis.}
}
\value{
A ggplot2 object for a set of boxplots of scores per K clusters
to show the effect of clustering on MST inference.
}
\description{
This function plots a boxplot per K number of clusters to look at
the distribution of scores of MSTs inferred when coordinates are partitioned into K clusters.
This plot in part explains the choice of K used in the optimal MST for curve fitting.
}
