% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{boxplot_scoring}
\alias{boxplot_scoring}
\title{Boxplots to visualize score distribution}
\usage{
boxplot_scoring(retro_obj)
}
\arguments{
\item{retro_obj}{RETRO meta object.}
}
\value{
A ggplot2 object for a set of boxplots of scores per K clusters
to show the effect of clustering on MST inference.
}
\description{
This function plots a boxplot per K number of clusters to look at
the distribution of scores of MSTs inferred when coordinates are partitioned into K clusters.
This plot in part explains the choice of K used in the optimal MST for curve fitting.
}
