% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro_lineages.R
\name{get_bezier_curve}
\alias{get_bezier_curve}
\title{Bézier spline fitting function}
\usage{
get_bezier_curve(retro_obj, extension = 2)
}
\arguments{
\item{retro_obj}{List returned by RETRO \code{create_meta_obj()} containing information
about the optimal MST after scoring.}

\item{extension}{(optional) Numerical parameter for extending the curve beyond the
trajectory starting and terminal nodes. Default is 2.}
}
\value{
List of curve information stored in \code{retro_obj@RETRO_Curve}.
The list elements are: (1) smoothed Bézier curves
\if{html}{\out{<br>}}(2) MST centroids used to determine the curve
\if{html}{\out{<br>}}(3) smoothed Bézier curves prior to extension
\if{html}{\out{<br>}}(4) the vector of arc lengths between each set of centroids.
}
\description{
This function creates the Bézier spline curve to smooth out the
piecewise MST structure inferred by RETRO. It achieves this by stitching together cubic Bézier curves,
where the initial and final control points are from the nodes of the underlying MST (cluster "centroids").
The remaining two control points are computed internally and are not outputted by RETRO.
}
