% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro_lineages.R
\name{get_num_lineages}
\alias{get_num_lineages}
\title{Determine consensus lineages in data.}
\usage{
get_num_lineages(retro_obj, percent = 0.05, cutoff = 0.8)
}
\arguments{
\item{retro_obj}{RETRO meta object}

\item{percent}{(optional) Fraction to select the top x\% of high-performing MSTs based on score. Default is 0.05.}

\item{cutoff}{(optional) Membership cutoff for whether a cell can be assigned to a lineage based
on its mapping to that lineage across inferred MSTs. Default is 0.8 (cell must be mapped to lineage in ≥80\% of selected MSTs)}
}
\value{
This function returns \code{retro_obj} with updated information regarding the optimal MST for pseudotime inference, as well
as well as membership lists separating cells into their respective developmental lineages.
}
\description{
Following the iterative MST inference across K clusters, this function
determines the optimal number of lineages based on the MST score and overlap
in paths that are detected by all MSTs.
}
