% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chadwick.R
\name{install_chadwick}
\alias{install_chadwick}
\title{Install Chadwick}
\usage{
install_chadwick(obj, ...)
}
\arguments{
\item{obj}{An \code{etl_retro} object}

\item{...}{currectly ignored}
}
\description{
Install Chadwick
}
