% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro_distance.R
\name{pc_distance_function}
\alias{pc_distance_function}
\title{Defining cell coordinates using gene expression and time}
\usage{
pc_distance_function(time, dimred, lambda = 1)
}
\arguments{
\item{time}{Vector of sampling time points corresponding to each cell.}

\item{dimred}{Dimensionality reduction data (ideally from \code{prcomp()}).}

\item{lambda}{A numerical value indicating contribution from PCA for scaling the \code{time} vector.
Default value is 1 (where PCA and time have an equal contribution). Increasing lambda will increase contribution from PCA.}
}
\value{
List containing weighted coordinates, PCA coordinates weighted by their variance,
and the computed weight used to scale the time vector.
}
\description{
This function concatenates weighted time points to PCA coordinates.
}
