% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro_pseudotime_est.R
\name{pseudotime_fit}
\alias{pseudotime_fit}
\title{Pseudotime fitting function}
\usage{
pseudotime_fit(retro_obj)
}
\arguments{
\item{retro_obj}{MST-specific object from RETRO containing information about the dataset
and the optimal MST used in curve fitting.}
}
\value{
Updated elements in \code{retro_obj} to store RETRO pseudotime in \code{pseudotime}
and intermediate fitting processes in \code{pseudotime_graph}.
}
\description{
This function uses lineage information from the optimal MST
to infer pseudotime values.
}
