% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro_clustering.R
\name{recluster}
\alias{recluster}
\title{Re-cluster cells given a period}
\usage{
recluster(kmnn, time, coordinates, period)
}
\arguments{
\item{kmnn}{Initial cluster centroids and labels from \code{kmnn_cluster()}.}

\item{period}{Specifies the minimum difference in time where one can expect the cells
to return to an earlier state gene expression (make a cycle).}
}
\value{
A list containing an updated membership vector \code{clusterLabels} following re-clustering,
the average difference in time between cycles, \code{cycle_dt}, and the \code{period}.
}
\description{
This function re-clusters the cells if there is periodicity in their gene expression,
causing clusters to contain multiple time points.
}
