% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.R
\name{reps}
\alias{reps}
\title{Replicate Elements of Vectors}
\usage{
reps(s, n)
}
\arguments{
\item{s}{vector to replicate}

\item{n}{number of times to repeat}
}
\description{
Replicate elements of vectors and concatenate the result into a single string.
}
\keyword{internal}

