% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro_scoring.R
\name{scoring}
\alias{scoring}
\title{Parallelization of MST generation and scoring}
\usage{
scoring(retro_obj, k_range, num_scores = 100)
}
\arguments{
\item{retro_obj}{RETRO meta object.}

\item{k_range}{Numeric vector for the range of K clusters that will be used to partition the coordinates}

\item{num_scores}{Number of MSTs (scored MSTs) per clustering. Default is 100.}
}
\value{
Lists in \code{retro_obj@all_scores} containing MST scoring information from \code{av_cell_dist()}
and in \code{retro_obj@k} for storing results from iterative clustering.
}
\description{
This function is a wrapper for the \code{generate_paths()} function
to parallelize iterative MST generation and scoring across the range of specified K clusters.
It is recommended to run \code{scoring()} in an lapply() to contain all
iterative clustering and scoring data in one .rda file.
}
